添加注释

+++ handler/src/test/java/io/netty/handler/stream/ChunkedWriteHandlerTest.java
@@ -225,7 +225,7 @@ public class ChunkedWriteHandlerTest {
         ch.writeAndFlush(input).syncUninterruptibly();
         assertTrue(ch.finish());
 
-        assertEquals(0, ch.readOutbound());
+        assertEquals(0, (long) ch.readOutbound());
         assertNull(ch.readOutbound());
     }
 
@@ -257,13 +257,13 @@ public class ChunkedWriteHandlerTest {
     @Test
     public void testSkipAfterFailedChunkedStream() throws IOException {
         checkSkipFailed(new ChunkedStream(new ByteArrayInputStream(BYTES)),
-                        new ChunkedStream(new ByteArrayInputStream(BYTES)));
+                new ChunkedStream(new ByteArrayInputStream(BYTES)));
     }
 
     @Test
     public void testSkipAfterFailedChunkedNioStream() throws IOException {
         checkSkipFailed(new ChunkedNioStream(Channels.newChannel(new ByteArrayInputStream(BYTES))),
-                        new ChunkedNioStream(Channels.newChannel(new ByteArrayInputStream(BYTES))));
+                new ChunkedNioStream(Channels.newChannel(new ByteArrayInputStream(BYTES))));
     }
 
     @Test
@@ -302,7 +302,7 @@ public class ChunkedWriteHandlerTest {
 
         // 3 out of 4 chunks were already written
         int read = 0;
-        for (;;) {
+        for (; ; ) {
             ByteBuf buffer = ch.readOutbound();
             if (buffer == null) {
                 break;
@@ -574,7 +574,7 @@ public class ChunkedWriteHandlerTest {
     private static void check(Object... inputs) {
         EmbeddedChannel ch = new EmbeddedChannel(new ChunkedWriteHandler());
 
-        for (Object input: inputs) {
+        for (Object input : inputs) {
             ch.writeOutbound(input);
         }
 
@@ -582,7 +582,7 @@ public class ChunkedWriteHandlerTest {
 
         int i = 0;
         int read = 0;
-        for (;;) {
+        for (; ; ) {
             ByteBuf buffer = ch.readOutbound();
             if (buffer == null) {
                 break;
@@ -645,7 +645,7 @@ public class ChunkedWriteHandlerTest {
         // we expect to see the second message, chunk by chunk
         int i = 0;
         int read = 0;
-        for (;;) {
+        for (; ; ) {
             ByteBuf buffer = ch.readOutbound();
             if (buffer == null) {
                 break;
